// frontend_temp/src/types/global.d.ts
export interface DialogBoxProps<InputData> {
	title?: string;
	description?: string;
	open?: boolean;
	setOpen?: (open: boolean) => !open;
	data?: InputData;
}

export interface ProgrammeType {
	_id: string;
	name: string;
}